Fix quantity appearance bug
quantity was showing as float, not as int, change type and method in ProductCursorAdapter

+++ app/src/main/java/com/ratajczykdev/inventoryapp/ProductCursorAdapter.java
@@ -73,7 +73,7 @@ public class ProductCursorAdapter extends CursorAdapter
         //  convert byte array to Bitmap
         Bitmap photo = BitmapFactory.decodeByteArray(byteArrayPhoto, 0, byteArrayPhoto.length);
         float price = cursor.getFloat(priceColumnIndex);
-        float quantity = cursor.getFloat(quantityColumnIndex);
+        int quantity = cursor.getInt(quantityColumnIndex);
 
         textName.setText(name);
         imagePhoto.setImageBitmap(photo);
